% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{add_tn_logo}
\alias{add_tn_logo}
\title{Add a TN logo to a plot}
\usage{
add_tn_logo(
  plot,
  logo,
  size = 0.1,
  position = "bottom left",
  margin = 1,
  vjust = 1
)
}
\arguments{
\item{plot}{ggplot object to add on to}

\item{logo}{Name of logo from tn_logo_names}

\item{size}{Vertical size of logo in relation to entire plot. size=.1 means the logo will take up a tenth of the vertical space}

\item{position}{String specifying top/bottom and left/center/right position of logo}

\item{margin}{Space around entire output in mm}

\item{vjust}{Vertical justification 0-1. 0 places the logo entirely inside the plot. 1 places the logo entirely outside.}
}
\description{
Add a TN logo to a plot
}
