% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{scale_color_tn}
\alias{scale_color_tn}
\title{Color scale constructor for TN colors}
\usage{
scale_color_tn(
  palette = "Default",
  discrete = TRUE,
  reverse = FALSE,
  interpolate = FALSE,
  ...
)
}
\arguments{
\item{palette}{Character name of palette in tn_palettes}

\item{discrete}{Boolean indicating whether color aesthetic is discrete or not}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{interpolate}{Boolean indicating whether colors should be interpolated from the palette if the color aesthetic is discrete}

\item{...}{Additional arguments passed to discrete_scale or
scale_color_gradientn, used respectively when discrete is TRUE or FALSE}
}
\description{
Color scale constructor for TN colors
}
