% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{theme_tn}
\alias{theme_tn}
\title{Set theme to comply with TN branding with options for coloring plot elements. Gives warnings on non-approved colors}
\usage{
theme_tn(
  title_color = NA,
  subtitle_color = NA,
  axis_text_color = NA,
  axis_line_color = NA,
  caption_color = NA,
  background_color = NA
)
}
\arguments{
\item{title_color}{Color name, TN color name, or hex. Will warn if not an approved color.}

\item{subtitle_color}{Color name, TN color name, or hex. Will warn if not an approved color.}

\item{axis_text_color}{Color name, TN color name, or hex. Will warn if not an approved color.}

\item{axis_line_color}{Color name, TN color name, or hex. Will warn if not an approved color.}

\item{caption_color}{Color name, TN color name, or hex. Will warn if not an approved color.}

\item{background_color}{Color name, TN color name, or hex. Will warn if not an approved color.}
}
\description{
Set theme to comply with TN branding with options for coloring plot elements. Gives warnings on non-approved colors
}
