% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{tn_color_show}
\alias{tn_color_show}
\title{Show all branding approved colors}
\usage{
tn_color_show(palette = NA, labels = T, duplicates = T)
}
\arguments{
\item{palette}{Optional filter to limit colors to certain palettes}

\item{labels}{Boolean indicating whether color info should be shown}

\item{duplicates}{Boolean indicating whether to show each color once, even if it is in multiple palettes}
}
\description{
Show all branding approved colors
}
