% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframes.R
\docType{data}
\name{tn_colors}
\alias{tn_colors}
\title{Branding colors for TN}
\format{
An object of class \code{data.frame} with 120 rows and 5 columns.
}
\usage{
tn_colors
}
\description{
Hex values, palette, family, and name info for branding approved colors
}
\keyword{datasets}
