% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframes.R
\docType{data}
\name{tn_logo_name_list}
\alias{tn_logo_name_list}
\title{Names associated with TN logos}
\format{
An object of class \code{character} of length 30.
}
\usage{
tn_logo_name_list
}
\description{
Names associated with TN logos
}
\keyword{datasets}
