% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{tn_pal}
\alias{tn_pal}
\title{Return function to interpolate a TN color palette}
\usage{
tn_pal(palette = "Primary", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in tn_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Optional arguments passed on to colorRampPalette}
}
\description{
Return function to interpolate a TN color palette
}
