% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{tn_palette_show}
\alias{tn_palette_show}
\title{Show all branding approved colors organized by palette}
\usage{
tn_palette_show(palette = NA)
}
\arguments{
\item{palette}{Optional filter to limit colors to certain palettes}

\item{labels}{Boolean indicating whether color info should be shown}

\item{duplicates}{Boolean indicating whether to show each color once, even if it is in multiple palettes}
}
\description{
Show all branding approved colors organized by palette
}
