% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframes.R
\docType{data}
\name{tn_palettes}
\alias{tn_palettes}
\title{Color palettes for TN}
\format{
An object of class \code{list} of length 19.
}
\usage{
tn_palettes
}
\description{
Color palettes for TN
}
\keyword{datasets}
