% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{tn_validate_color}
\alias{tn_validate_color}
\title{Check that a color is branding compliant}
\usage{
tn_validate_color(color)
}
\arguments{
\item{color}{Color string, hex, or TN name}
}
\value{
Hex color if valid TN color, original input otherwise
}
\description{
Check that a color is branding compliant
}
